FIX Fixed the capture tests and removed a bug in the captures
This was a bit of a complicated fix. See, the teardown_method() function in the test_captures.py file really was not meant to be used in combination with fixtures. Fixtures already have ways to do their own cleanup, and they are self-contained. The teardown was being run after EVERY method call. This lead to some wonky behaviour.

Not only ImageCapture was broken: actually every capture was not really working as intended when you read the next frame. I've added some code so that the get_next_frame returns false if you closed the capture beforehand.

The fix was to change the scope of the fixture to "class". Instead of every test being run with every fixture in sequence, every test is now being run in sequence with the same fixture. Even if a test fails, the "test_closed" will now always be run (and if THAT fails you have other problems).


Former-commit-id: 56a4fa51baeb97ac965ac426050cbe8b97d24abc