FFT Handle lost forwarder
When the stream stops working the player will wait for a set amount of time.
After this time runs out a message will pop up as an overlay on the videoplayer telling the user to check the forwarder.
When closing the message the player will attempt a reload without refreshing the entire page